Fix torch fit
